+++ tasks/index.js
@@ -1,3 +1,4 @@
 
 require('./deploy.js')
+require('./update.js')
 require('./try-sign.js')
